Added fadmin spectp
- /fadmin spectp
Allows admins to spectate a claim in spectator mode or be returned to their original location back in survival. Added button for this to /fadmin devrequests
Players are automatically returned when they leave the server or the server is shut down; however in the event of a crash the data is not saved on file and may require manual intervention to return players to their correct locations.